# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for digital queue",
        "title": "Digital Queue API",
        "contact": {
            "name": "Timur Aliev",
            "url": "https://t.me/Aliev_Timur_M",
            "email": "alievtm@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/logout": {
            "post": {
                "description": "logout account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "operationId": "logout",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.logoutInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/types.LogoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "refresh AccessToken, RefreshToken, Employee, Workstation data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "operationId": "refresh",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.refreshTokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/types.AuthorizationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "employee authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign In",
                "operationId": "sign-in",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.signInInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/types.AuthorizationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "registration new account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign Up",
                "operationId": "registration-account",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee ID",
                        "schema": {
                            "$ref": "#/definitions/types.SignUpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.logoutInput": {
            "type": "object",
            "required": [
                "employeeId"
            ],
            "properties": {
                "employeeId": {
                    "type": "string"
                }
            }
        },
        "handler.refreshTokenInput": {
            "type": "object",
            "required": [
                "employeeId",
                "refreshToken",
                "workstationId"
            ],
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "workstationId": {
                    "type": "string"
                }
            }
        },
        "handler.signInInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.AuthorizationResponse": {
            "type": "object",
            "required": [
                "accessToken",
                "employee",
                "refreshToken",
                "workstation"
            ],
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "employee": {
                    "$ref": "#/definitions/types.Employee"
                },
                "refreshToken": {
                    "type": "string"
                },
                "workstation": {
                    "$ref": "#/definitions/types.Workstation"
                }
            }
        },
...
        "types.Employee": {
            "type": "object",
            "properties": {
                "employee_id": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "second_name": {
                    "type": "string"
                },
                "session_state": {
                    "type": "boolean"
                },
                "status": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "workstation_id": {
                    "type": "integer"
                }
            }
        },
...
        "types.LogoutResponse": {
            "type": "object",
            "required": [
                "status_response"
            ],
            "properties": {
                "status_response": {
                    "type": "boolean"
                }
            }
        },
...
        "types.SignUpResponse": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "types.Workstation": {
            "type": "object",
            "properties": {
                "employee_id": {
                    "type": "integer"
                },
                "workstation_id": {
                    "type": "integer"
                },
                "workstation_name": {
                    "type": "string"
                }
            }
        }
    }
}